chineseocr/requirements.txt
scipy
numpy
jupyter
ipython
easydict
opencv-contrib-python
Cython
pillow
h5py
flask
flask_restful
lmdb
mahotas
pandas
requests
web.py==0.40.dev0
redis
pika
matplotlib
lxml
bs4
regex
pdf2image
python-docx